microservices/user-service/requirements.txt
# ==========================================
# DJANGO CORE
# ==========================================
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# ==========================================
# DATABASE & ORM
# ==========================================
psycopg2-binary==2.9.7
prisma==0.10.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.22

# ==========================================
# AUTHENTICATION & SECURITY
# ==========================================
PyJWT==2.8.0
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
bcrypt==4.1.2
argon2-cffi==23.1.0

# ==========================================
# CACHE & QUEUE
# ==========================================
redis==5.0.1
celery==5.3.4
django-redis==5.3.0
django-celery-results==2.5.1

# ==========================================
# CONFIGURATION & ENVIRONMENT
# ==========================================
python-decouple==3.8
dj-database-url==2.1.0

# ==========================================
# API & HTTP CLIENTS
# ==========================================
requests==2.31.0
httpx==0.25.2
drf-yasg==1.21.5

# ==========================================
# PRODUCTION SERVER
# ==========================================
gunicorn==21.2.0
whitenoise==6.6.0

# ==========================================
# FILE STORAGE
# ==========================================
Pillow==10.1.0

# ==========================================
# DATA VALIDATION & SERIALIZATION
# ==========================================
pydantic==2.5.0

# ==========================================
# DEVELOPMENT & CODE QUALITY
# ==========================================
black==23.11.0
isort==5.13.2
flake8==6.1.0
pylint-django==2.5.5
django-extensions==3.2.3

# ==========================================
# TESTING
# ==========================================
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1
factory-boy==3.3.0
freezegun==1.4.0
model-bakery==1.15.0
Faker==21.0.1

# ==========================================
# DEBUGGING
# ==========================================
django-debug-toolbar==4.2.0
ipython==8.17.2

# ==========================================
# MONITORING & LOGGING
# ==========================================
structlog==23.2.0
django-structlog==4.0.0
sentry-sdk==1.35.0

# ==========================================
# UTILITIES
# ==========================================
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7